📝 chore: match/detail 하위 디렉터리 구조 수정
