+++ src/screens/match/detail/index.ts
@@ -1,5 +1,5 @@
 export * from './MatchDetailScreen';
 export * from './MatchDetailProfileScreen';
-export * from './MatchDetailRecordScreen';
 export * from './MatchDetailMatchingScreen';
 export * from './MatchDetailMemberScreen';
+export * from './record';
