+++ src/screens/match/detail/record/index.ts
@@ -1,0 +1,1 @@
+export * from './MatchDetailRecordScreen';
